Added shopp('checkout','billing-required') tag to help hide payment fields when the order is free Added category pagination CSS placeholders in shopp.css Added support for different line breaks in the email template system Added support for special category-specific and product-specific template files (category-id.php and product-id.php) Added custom data input support for products using the shopp('product','input') tag Enhanced checkout process to bypass transaction engine when the order is free Fixed an issue with thumbnail image scale settings not being applied correctly Added support for handling WordPress logouts when using WordPress integrated accounts Fixed character escape issues in category variation option template labels Fixes for the Bestseller dashboard widget Added support for live shipping rate modules Fixed an issue with generating the Google API callback URL Cleaned up admin styles for the order manager and for the shipping rates tables Fixed the default AJAX cart handler to better handle items without thumbnail images Fixed an issue with multiple variation menus when no defaults option is set Fixed an issue to ensure that IE correctly handles the same shipping address toggle in the checkout form Fix for an issue with adding multiple tax rates for different countries without any state/province regions Enhancements to the Shopp DB wrapper and database schema to be compatible with MySQL strict mode Changed the UK unit system to metric Corrected Slovenia's currency to Euros Fixed image looping in the shopp('product') tags Added options to the gallery for thumbnail sizes (thumbsize, thumbheight, thumbwidth) Fixed an issue with product variation menus in multi-mode when the product is repeatedly displayed on a single page under different categories Fixed an issue with file system storage that allows multiple assets to be tied to a single file and removes the file when all references to it are deleted Added shopp('subcategory') tag support Enhanced category pagination (from Robbert Langezaal) Properly closing nested category lists in the category-list tag/widget Optimized cart totaling calculations Fixed a bad PHP tag escape sequence in the PayPal Pro addon Added custom field and pagestyle field support to PayPal Express Fixes for Google Checkout to ensure Shopp only handles Shopp-originating orders Fix to ensure Google uses the locale-bsaed currency first and gracefully degrades to use USD Fixes to the Authorize.net addon settings Added UPS live shipping rates addon Added FedEx live shipping rate addon Updated i18n files Added Dutch language support Fixed issues with WordPress integrated account login support Fixed the Shopp browser titles Adjust the confirm order redirect to use secure connections Fix for detecting secure requests (https) more accurately Fixes for numerous links when pretty permalinks are disabled
git-svn-id: file:///Users/jond/svngit/svnrepos/shopp/trunk@326 846ab242-8549-0410-9a75-dfdbdae9473e

+++ core/model/Asset.php
@@ -85,20 +85,25 @@ class Asset extends DatabaseObject {
 
 		$selection = "";
 		foreach ($keys as $value) 
-			$selection .= ((!empty($selection))?" OR ":"")."{$this->_key}=$value OR src=$value";
+			$selection .= ((!empty($selection))?" OR ":"")."f.{$this->_key}=$value OR f.src=$value";
 
 		if ($this->storage == "fs") $this->deletefiles($selection);
 
-		$query = "DELETE LOW_PRIORITY FROM $this->_table WHERE $selection";
+		$query = "DELETE LOW_PRIORITY FROM $this->_table AS f WHERE $selection";
 		$db->query($query);
 	}
-	
+
+	/** 
+	 * deletefiles ()
+	 * Remove files from the file system only when 1 reference to the file exists
+	 * in file references in the database, otherwise, leave them **/
 	function deletefiles ($selection) {
 		$db =& DB::get();
 		
-		$files = $db->query("SELECT name FROM $this->_table WHERE $selection",AS_ARRAY);
-		foreach ($files as $file) 
-			unlink($this->path.$file->name);
+		$files = $db->query("SELECT f.name,count(DISTINCT links.id) AS refs FROM $this->_table AS f LEFT JOIN $this->_table AS links ON f.name=links.name WHERE $selection GROUP BY links.name",AS_ARRAY);
+		foreach ($files as $file)
+			if ($file->refs == 1 && file_exists($this->path.$file->name))
+				unlink($this->path.$file->name);
 
 		return true;
 	}
